refactor flutter service
