--- sys-account/service/flu/lib/core/core.dart
@@ -1,4 +1,0 @@
-export './routes/paths.dart';
-export './shared_services/base_model.dart';
-export './shared_repositories/auth_repo.dart';
-export './shared_repositories/account_repo.dart';
