--- sys-account/service/flu/lib/modules/auth_dialog.dart
@@ -1,3 +1,0 @@
-library auth_dialog;
-
-export 'account_module.dart';
