+++ sys-account/service/flu/lib/pkg/shared_repositories/auth_repo.dart
@@ -1,4 +1,4 @@
-import 'package:auth_dialog/core/shared_repositories/base_repo.dart';
+import 'package:sys_share_sys_account_service/logic/shared_repositories/base_repo.dart';
 import 'package:sys_share_sys_account_service/sys_share_sys_account_service.dart'
     as rpc;
 import 'package:meta/meta.dart';
