+++ sys-account/service/flu/lib/view/view.dart
@@ -1,0 +1,2 @@
+export 'widgets/auth_dialog.dart';
+export 'screens/account_view.dart';
